Implements user login and logout
Adds user login and logout functionality to the application.

This commit introduces the ability for users to log in and out
of the application.  It replaces the previous "select user"
functionality with a more realistic login simulation.  The
UI now displays information about the currently logged-in user
and provides a logout button.  The "Add Todo" form is disabled
until a user is logged in, improving the user experience.