Refactors todo ownership to use general maps
Refactors the todo ownership model to utilize the general maps package for managing relationships between users and todos.

This change removes the direct `user_id` foreign key from the `todos` table and introduces new endpoints for assigning and removing user associations with todos. The `owner()` method is added to the `Todo` model to retrieve the todo's owner. The update method is modified to allow user reassignment.

This approach offers more flexibility and scalability for managing relationships and metadata in the application.